Refactor order class
