+++ src/main/cl/uchile/dcc/neural/networks/service/PerceptonBuilder.java
@@ -1,17 +1,17 @@
-package cl.uchile.dcc.neural.networks;
+package main.cl.uchile.dcc.neural.networks.service;
 
 import java.util.Arrays;
 import java.util.List;
 
 public class PerceptonBuilder {
 
-    List<Double> nandWeights = Arrays.asList(-2.0, -2.0);
-    List<Double> orWeights = Arrays.asList(7.0, 7.0);
-    List<Double> andWeights = Arrays.asList(2.0, 2.0);
+    private List<Double> nandWeights = Arrays.asList(-2.0, -2.0);
+    private List<Double> orWeights = Arrays.asList(7.0, 7.0);
+    private List<Double> andWeights = Arrays.asList(2.0, 2.0);
 
-    int nandBias = 1;
-    int orBias = -5;
-    int andBias = -3;
+    private int nandBias = 4;
+    private int orBias = -5;
+    private int andBias = -3;
 
     public Perceptron getNandPerceptron(){
         return new Perceptron(nandWeights, nandBias);
